+++ plugins/MoleculeExplorer/io/protein/ProteinLoader.h
@@ -21,28 +21,26 @@
 #pragma once
 
 #include <brayns/io/Loader.h>
-#include <brayns/io/loaders/protein/ProteinLoaderParameters.h>
 
-namespace brayns
-{
+#include "ProteinLoaderParameters.h"
+
 /**
  * @brief The ProteinLoader class loads protein from PDB files: http://www.rcsb.org
  */
-class ProteinLoader : public Loader<ProteinLoaderParameters>
+class ProteinLoader : public brayns::Loader<ProteinLoaderParameters>
 {
 public:
     std::vector<std::string> getSupportedExtensions() const final;
 
     std::string getName() const final;
 
-    std::vector<std::shared_ptr<Model>> importFromFile(
+    std::vector<std::shared_ptr<brayns::Model>> importFromFile(
         const std::string &path,
-        const LoaderProgress &callback,
+        const brayns::LoaderProgress &callback,
         const ProteinLoaderParameters &parameters) const final;
 
-    std::vector<std::shared_ptr<Model>> importFromBlob(
-        const Blob &blob,
-        const LoaderProgress &callback,
+    std::vector<std::shared_ptr<brayns::Model>> importFromBlob(
+        const brayns::Blob &blob,
+        const brayns::LoaderProgress &callback,
         const ProteinLoaderParameters &parameters) const final;
 };
-} // namespace brayns